Crate inkwell

Inkwell documentation is a work in progress.

If you have any LLVM knowledge that could be used to improve these docs, we would greatly appreciate you opening an issue and/or a pull request on our GitHub page.

Due to a rustdoc issue, this documentation represents only the latest supported LLVM version. We hope that this issue will be resolved in the future so that multiple versions can be documented side by side.

Library Wide Notes

Most functions which take a string slice as input may possibly panic in the unlikely event that a c style string cannot be created based on it. (IE if your slice already has a null byte in it)

Re-exports

pub extern crate llvm_sys_170 as llvm_sys;

Modules

attributes

Attributes are optional modifiers to functions, function parameters, and return types.

basic_block

A BasicBlock is a container of instructions.

builder

A Builder enables you to build instructions.

comdat

A Comdat helps resolve linker errors for duplicate sections.

context

A Context is an opaque owner and manager of core global data.

data_layout

debug_info

Debug symbols - DebugInfoBuilder interface

execution_engine

intrinsics

memory_buffer

module

A Module represents a single code compilation unit.

object_file

passes

support

targets

types

A type is a classification which determines how data is used.

values

A value is an instance of a type.

Structs

AddressSpace

Defines the address space in which a global will be inserted.

Enums

AtomicOrdering

AtomicRMWBinOp

DLLStorageClass

FloatPredicate

Defines how to compare a left and right FloatValue.

GlobalVisibility

InlineAsmDialectllvm7-0 or llvm8-0 or llvm9-0 or llvm10-0 or llvm11-0 or llvm12-0 or llvm13-0 or llvm14-0 or llvm15-0 or llvm16-0 or llvm17-0

IntPredicate

This enum defines how to compare a left and right IntValue.

OptimizationLevel

Defines the optimization level used to compile a Module.

ThreadLocalMode