Struct inkwell::attributes::Attribute

pub struct Attribute { /* fields omitted */ }

Functions, function parameters, and return types can have Attributes to indicate how they should be treated by optimizations and code generation.

Implementations

impl Attribute

pub fn is_enum(self) -> bool

Determines whether or not an Attribute is an enum. This method will likely be removed in the future in favor of Attributes being generically defined.

Example

use inkwell::context::Context;

let context = Context::create();
let enum_attribute = context.create_enum_attribute(0, 10);

assert!(enum_attribute.is_enum());

pub fn is_string(self) -> bool

Determines whether or not an Attribute is a string. This method will likely be removed in the future in favor of Attributes being generically defined.

Example

use inkwell::context::Context;

let context = Context::create();
let string_attribute = context.create_string_attribute("my_key_123", "my_val");

assert!(string_attribute.is_string());

pub fn get_named_enum_kind_id(name: &str) -> u32

Gets the enum kind id associated with a builtin name.

Example

use inkwell::attributes::Attribute;

// This kind id doesn't exist:
assert_eq!(Attribute::get_named_enum_kind_id("foobar"), 0);

// These are real kind ids:
assert_eq!(Attribute::get_named_enum_kind_id("align"), 1);
assert_eq!(Attribute::get_named_enum_kind_id("builtin"), 5);

pub fn get_enum_kind_id(self) -> u32

Gets the kind id associated with an enum Attribute.

Example

use inkwell::context::Context;

let context = Context::create();
let enum_attribute = context.create_enum_attribute(0, 10);

assert_eq!(enum_attribute.get_enum_kind_id(), 0);

pub fn get_last_enum_kind_id() -> u32

Gets the last enum kind id associated with builtin names.

Example

use inkwell::attributes::Attribute;

assert_eq!(Attribute::get_last_enum_kind_id(), 56);

pub fn get_enum_value(self) -> u64

Gets the value associated with an enum Attribute.

Example

use inkwell::context::Context;

let context = Context::create();
let enum_attribute = context.create_enum_attribute(0, 10);

assert_eq!(enum_attribute.get_enum_value(), 10);

pub fn get_string_kind_id(&self) -> &CStr

Gets the string kind id associated with a string attribute.

Example

use inkwell::context::Context;

let context = Context::create();
let string_attribute = context.create_string_attribute("my_key", "my_val");

assert_eq!(string_attribute.get_string_kind_id().to_str(), Ok("my_key"));

pub fn get_string_value(&self) -> &CStr

Gets the string value associated with a string attribute.

Example

use inkwell::context::Context;

let context = Context::create();
let string_attribute = context.create_string_attribute("my_key", "my_val");

assert_eq!(string_attribute.get_string_value().to_str(), Ok("my_val"));