Struct inkwell::builder::Builder

pub struct Builder { /* fields omitted */ }

Methods

impl Builder

pub fn create() -> Self

pub fn build_return(&self, value: Option<&BasicValue>) -> InstructionValue

pub fn build_aggregate_return(
&self,
values: &[BasicValueEnum]
) -> InstructionValue

Important traits for Either<L, R>
impl<L, R> Write for Either<L, R> where
L: Write,
R: Write, impl<L, R> Iterator for Either<L, R> where
L: Iterator,
R: Iterator<Item = <L as Iterator>::Item>, type Item = <L as Iterator>::Item;impl<L, R> Read for Either<L, R> where
L: Read,
R: Read,

pub fn build_call(
&self,
function: FunctionValue,
args: &[BasicValueEnum],
name: &str,
tail_call: bool
) -> Either<BasicValueEnum, InstructionValue>

pub unsafe fn build_gep(
&self,
ptr: PointerValue,
ordered_indexes: &[IntValue],
name: &str
) -> PointerValue

GEP is very likely to segfault if indexes are used incorrectly, and is therefore an unsafe function. Maybe we can change this in the future.

pub unsafe fn build_in_bounds_gep(
&self,
ptr: PointerValue,
ordered_indexes: &[IntValue],
name: &str
) -> PointerValue

GEP is very likely to segfault if indexes are used incorrectly, and is therefore an unsafe function. Maybe we can change this in the future.

pub unsafe fn build_struct_gep(
&self,
ptr: PointerValue,
index: u32,
name: &str
) -> PointerValue

GEP is very likely to segfault if indexes are used incorrectly, and is therefore an unsafe function. Maybe we can change this in the future.

pub fn build_ptr_diff(
&self,
lhs_ptr: PointerValue,
rhs_ptr: PointerValue,
name: &str
) -> IntValue

pub fn build_phi<T: BasicType>(&self, type_: T, name: &str) -> PhiValue

pub fn build_store<V: BasicValue>(
&self,
ptr: PointerValue,
value: V
) -> InstructionValue

pub fn build_load(&self, ptr: PointerValue, name: &str) -> BasicValueEnum

pub fn build_alloca<T: BasicType>(&self, ty: T, name: &str) -> PointerValue

pub fn build_array_alloca<T: BasicType>(
&self,
ty: T,
size: IntValue,
name: &str
) -> PointerValue

pub fn build_malloc<T: BasicType>(&self, ty: T, name: &str) -> PointerValue

pub fn build_array_malloc<T: BasicType>(
&self,
ty: T,
size: IntValue,
name: &str
) -> PointerValue

pub fn build_free(&self, ptr: PointerValue) -> InstructionValue

pub fn insert_instruction(
&self,
instruction: &InstructionValue,
name: Option<&str>
)

pub fn get_insert_block(&self) -> Option<BasicBlock>

pub fn build_int_unsigned_div<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_signed_div<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_exact_signed_div<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_unsigned_rem<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_signed_rem<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_s_extend<T: IntMathValue>(
&self,
int_value: T,
int_type: T::BaseType,
name: &str
) -> T

pub fn build_int_s_extend_or_bit_cast<T: IntMathValue>(
&self,
int_value: T,
int_type: T::BaseType,
name: &str
) -> T

pub fn build_int_z_extend<T: IntMathValue>(
&self,
int_value: T,
int_type: T::BaseType,
name: &str
) -> T

pub fn build_int_z_extend_or_bit_cast<T: IntMathValue>(
&self,
int_value: T,
int_type: T::BaseType,
name: &str
) -> T

pub fn build_int_truncate<T: IntMathValue>(
&self,
int_value: T,
int_type: T::BaseType,
name: &str
) -> T

pub fn build_int_truncate_or_bit_cast<T: IntMathValue>(
&self,
int_value: T,
int_type: T::BaseType,
name: &str
) -> T

pub fn build_float_rem<T: FloatMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_float_to_unsigned_int<T: FloatMathValue>(
&self,
float: T,
int_type: <T::BaseType as FloatMathType>::MathConvType,
name: &str
) -> <<T::BaseType as FloatMathType>::MathConvType as IntMathType>::ValueType

pub fn build_float_to_signed_int<T: FloatMathValue>(
&self,
float: T,
int_type: <T::BaseType as FloatMathType>::MathConvType,
name: &str
) -> <<T::BaseType as FloatMathType>::MathConvType as IntMathType>::ValueType

pub fn build_unsigned_int_to_float<T: IntMathValue>(
&self,
int: T,
float_type: <T::BaseType as IntMathType>::MathConvType,
name: &str
) -> <<T::BaseType as IntMathType>::MathConvType as FloatMathType>::ValueType

pub fn build_signed_int_to_float<T: IntMathValue>(
&self,
int: T,
float_type: <T::BaseType as IntMathType>::MathConvType,
name: &str
) -> <<T::BaseType as IntMathType>::MathConvType as FloatMathType>::ValueType

pub fn build_float_trunc<T: FloatMathValue>(
&self,
float: T,
float_type: T::BaseType,
name: &str
) -> T

pub fn build_float_ext<T: FloatMathValue>(
&self,
float: T,
float_type: T::BaseType,
name: &str
) -> T

pub fn build_float_cast<T: FloatMathValue>(
&self,
float: T,
float_type: T::BaseType,
name: &str
) -> T

pub fn build_int_cast<T: IntMathValue>(
&self,
int: T,
int_type: T::BaseType,
name: &str
) -> T

pub fn build_float_div<T: FloatMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_add<T: IntMathValue>(&self, lhs: T, rhs: T, name: &str) -> T

pub fn build_int_nsw_add<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_nuw_add<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_float_add<T: FloatMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_xor<T: IntMathValue>(&self, lhs: T, rhs: T, name: &str) -> T

pub fn build_and<T: IntMathValue>(&self, lhs: T, rhs: T, name: &str) -> T

pub fn build_or<T: IntMathValue>(&self, lhs: T, rhs: T, name: &str) -> T

pub fn build_left_shift<T: IntMathValue>(&self, lhs: T, rhs: T, name: &str) -> T

Builds an IntValue containing the result of a logical left shift instruction.

Example

A logical left shift is an operation in which an integer value's bits are shifted left by N number of positions.

assert_eq!(0b0000_0001 << 0, 0b0000_0001);
assert_eq!(0b0000_0001 << 1, 0b0000_0010);
assert_eq!(0b0000_0011 << 2, 0b0000_1100);

In Rust, a function that could do this for 8bit values looks like:

fn left_shift(value: u8, n: u8) -> u8 {
    value << n
}

And in Inkwell, the corresponding function would look roughly like:

use inkwell::context::Context;

// Setup
let context = Context::create();
let module = context.create_module("my_module");
let builder = context.create_builder();
let i8_type = context.i8_type();
let fn_type = i8_type.fn_type(&[i8_type.into(), i8_type.into()], false);

// Function Definition
let function = module.add_function("left_shift", &fn_type, None);
let value = function.get_first_param().unwrap().into_int_value();
let n = function.get_nth_param(1).unwrap().into_int_value();
let entry_block = function.append_basic_block("entry");

builder.position_at_end(&entry_block);

let shift = builder.build_left_shift(value, n, "left_shift"); // value << n

builder.build_return(Some(&shift));

pub fn build_right_shift<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
sign_extend: bool,
name: &str
) -> T

Builds an IntValue containing the result of a right shift instruction.

Example

A right shift is an operation in which an integer value's bits are shifted right by N number of positions. It may either be logical and have its leftmost N bit(s) filled with zeros or sign extended and filled with ones if the leftmost bit was one.

// Logical Right Shift
assert_eq!(0b1100_0000 >> 2, 0b0011_0000);
assert_eq!(0b0000_0010 >> 1, 0b0000_0001);
assert_eq!(0b0000_1100 >> 2, 0b0000_0011);

// Sign Extended Right Shift
assert_eq!(0b0100_0000i8 >> 2, 0b0001_0000);
assert_eq!(0b1110_0000i8 >> 1, 0b1111_0000);
assert_eq!(0b1100_0000i8 >> 2, 0b1111_0000);

In Rust, functions that could do this for 8bit values look like:

fn logical_right_shift(value: u8, n: u8) -> u8 {
    value >> n
}

fn sign_extended_right_shift(value: i8, n: u8) -> i8 {
    value >> n
}

Notice that, in Rust (and most other languages), whether or not a value is sign extended depends wholly on whether or not the type is signed (ie an i8 is a signed 8 bit value). LLVM does not make this distinction for you.

In Inkwell, the corresponding functions would look roughly like:

use inkwell::context::Context;

// Setup
let context = Context::create();
let module = context.create_module("my_module");
let builder = context.create_builder();
let i8_type = context.i8_type();
let fn_type = i8_type.fn_type(&[i8_type.into(), i8_type.into()], false);

// Function Definition
let function = module.add_function("right_shift", &fn_type, None);
let value = function.get_first_param().unwrap().into_int_value();
let n = function.get_nth_param(1).unwrap().into_int_value();
let entry_block = function.append_basic_block("entry");

builder.position_at_end(&entry_block);

// Whether or not your right shift is sign extended (true) or logical (false) depends
// on the boolean input parameter:
let shift = builder.build_right_shift(value, n, false, "right_shift"); // value >> n

builder.build_return(Some(&shift));

pub fn build_int_sub<T: IntMathValue>(&self, lhs: T, rhs: T, name: &str) -> T

pub fn build_int_nsw_sub<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_nuw_sub<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_float_sub<T: FloatMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_mul<T: IntMathValue>(&self, lhs: T, rhs: T, name: &str) -> T

pub fn build_int_nsw_mul<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_int_nuw_mul<T: IntMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_float_mul<T: FloatMathValue>(
&self,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_cast<T: BasicType, V: BasicValue>(
&self,
op: InstructionOpcode,
from_value: V,
to_type: T,
name: &str
) -> BasicValueEnum

pub fn build_pointer_cast<T: PointerMathValue>(
&self,
from: T,
to: T::BaseType,
name: &str
) -> T

pub fn build_int_compare<T: IntMathValue>(
&self,
op: IntPredicate,
lhs: T,
rhs: T,
name: &str
) -> T

pub fn build_float_compare<T: FloatMathValue>(
&self,
op: FloatPredicate,
lhs: T,
rhs: T,
name: &str
) -> <<T::BaseType as FloatMathType>::MathConvType as IntMathType>::ValueType

pub fn build_unconditional_branch(
&self,
destination_block: &BasicBlock
) -> InstructionValue

pub fn build_conditional_branch(
&self,
comparison: &IntValue,
then_block: &BasicBlock,
else_block: &BasicBlock
) -> InstructionValue

pub fn build_int_neg<T: IntMathValue>(&self, value: &T, name: &str) -> T

pub fn build_int_nsw_neg<T: IntMathValue>(&self, value: &T, name: &str) -> T

pub fn build_int_nuw_neg<T: IntMathValue>(&self, value: &T, name: &str) -> T

pub fn build_float_neg<T: FloatMathValue>(&self, value: &T, name: &str) -> T

pub fn build_not<T: IntMathValue>(&self, value: &T, name: &str) -> T

pub fn position_at(
&self,
basic_block: &BasicBlock,
instruction: &InstructionValue
)

pub fn position_before(&self, instruction: &InstructionValue)

pub fn position_at_end(&self, basic_block: &BasicBlock)

pub fn build_extract_value(
&self,
value: &AggregateValue,
index: u32,
name: &str
) -> BasicValueEnum

pub fn build_insert_value<V: BasicValue>(
&self,
value: V,
ptr: PointerValue,
index: u32,
name: &str
) -> InstructionValue

pub fn build_extract_element(
&self,
vector: VectorValue,
index: IntValue,
name: &str
) -> BasicValueEnum

pub fn build_insert_element<V: BasicValue>(
&self,
vector: VectorValue,
element: V,
index: IntValue,
name: &str
) -> BasicValueEnum

pub fn build_unreachable(&self) -> InstructionValue

pub fn build_fence(
&self,
atmoic_ordering: LLVMAtomicOrdering,
num: i32,
name: &str
) -> InstructionValue

pub fn build_is_null<T: PointerMathValue>(
&self,
ptr: T,
name: &str
) -> <<T::BaseType as PointerMathType>::PtrConvType as IntMathType>::ValueType

pub fn build_is_not_null<T: PointerMathValue>(
&self,
ptr: T,
name: &str
) -> <<T::BaseType as PointerMathType>::PtrConvType as IntMathType>::ValueType

pub fn build_int_to_ptr<T: IntMathValue>(
&self,
int: T,
ptr_type: <T::BaseType as IntMathType>::PtrConvType,
name: &str
) -> <<T::BaseType as IntMathType>::PtrConvType as PointerMathType>::ValueType

pub fn build_ptr_to_int<T: PointerMathValue>(
&self,
ptr: T,
int_type: <T::BaseType as PointerMathType>::PtrConvType,
name: &str
) -> <<T::BaseType as PointerMathType>::PtrConvType as IntMathType>::ValueType

pub fn clear_insertion_position(&self)

pub fn build_switch(
&self,
value: &IntValue,
else_block: &BasicBlock,
cases: &[(&IntValue, &BasicBlock)]
) -> InstructionValue

pub unsafe fn build_global_string(&self, value: &str, name: &str) -> GlobalValue

pub fn build_global_string_ptr(&self, value: &str, name: &str) -> GlobalValue