inkwell

Module context

A Context is an opaque owner and manager of core global data.

Structs

Context

A Context is a container for all LLVM entities including Modules.

ContextRef

A ContextRef is a smart pointer allowing borrowed access to a type’s Context.

Traits

AsContextRef

This trait abstracts an LLVM Context type and should be implemented with caution.