inkwell::passes
Struct PassBuilderOptions

pub struct PassBuilderOptions { /* private fields */ }

Available on crate features llvm13-0 or llvm14-0 or llvm15-0 or llvm16-0 or llvm17-0 or llvm18-0 only.

Implementations

impl PassBuilderOptions

pub fn create() -> Self

Create a new set of options for a PassBuilder

pub fn as_mut_ptr(&self) -> LLVMPassBuilderOptionsRef

Acquires the underlying raw pointer belonging to this PassBuilderOptions type.

pub fn set_verify_each(&self, value: bool)

Toggle adding the VerifierPass for the PassBuilder, ensuring all functions inside the module is valid.

pub fn set_debug_logging(&self, value: bool)

Toggle debug logging when running the PassBuilder.

pub fn set_loop_interleaving(&self, value: bool)

pub fn set_loop_vectorization(&self, value: bool)

pub fn set_loop_slp_vectorization(&self, value: bool)

pub fn set_loop_unrolling(&self, value: bool)

pub fn set_forget_all_scev_in_loop_unroll(&self, value: bool)

pub fn set_licm_mssa_opt_cap(&self, value: u32)

pub fn set_licm_mssa_no_acc_for_promotion_cap(&self, value: u32)

pub fn set_call_graph_profile(&self, value: bool)

pub fn set_merge_functions(&self, value: bool)