Trait inkwell::types::FloatMathType

pub unsafe trait FloatMathType<'ctx>: BasicType<'ctx> {
    type ValueType: FloatMathValue<'ctx>;
    type MathConvType: IntMathType<'ctx>;
}

Represents an LLVM type that can have floating point math operations applied to it.

Required Associated Types

type ValueType: FloatMathValue<'ctx>

The value instance of a float or float vector type.

type MathConvType: IntMathType<'ctx>

The type for float to int or float vector to int vector conversions.