Struct inkwell::values::CallSiteValue

pub struct CallSiteValue(_);

A value resulting from a function call. It may have function attributes applied to it.

This struct may be removed in the future in favor of an InstructionValue<CallSite> type.

Methods

impl CallSiteValue

pub fn set_tail_call(&self, tail_call: bool)

Sets whether or not this call is a tail call.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.set_tail_call(true);

pub fn is_tail_call(&self) -> bool

Determines whether or not this call is a tail call.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.set_tail_call(true);

assert!(call_site_value.is_tail_call());

pub fn try_as_basic_value(&self) -> Either<BasicValueEnum, InstructionValue>

Try to convert this CallSiteValue to a BasicValueEnum if not a void return type.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

assert!(call_site_value.try_as_basic_value().is_right());

pub fn add_attribute(&self, index: u32, attribute: Attribute)

Adds an Attribute to this CallSiteValue.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.add_attribute(0, string_attribute);
call_site_value.add_attribute(0, enum_attribute);

pub fn get_called_fn_value(&self) -> FunctionValue

Gets the FunctionValue this CallSiteValue is based on.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

assert_eq!(call_site_value.get_called_fn_value(), fn_value);

pub fn count_attributes(&self, index: u32) -> u32

Counts the number of Attributes on this CallSiteValue at an index.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.add_attribute(0, string_attribute);
call_site_value.add_attribute(0, enum_attribute);

assert_eq!(call_site_value.count_attributes(0), 2);

pub fn get_enum_attribute(&self, index: u32, kind_id: u32) -> Option<Attribute>

Gets an enum Attribute on this CallSiteValue at an index and kind id.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.add_attribute(0, string_attribute);
call_site_value.add_attribute(0, enum_attribute);

assert_eq!(call_site_value.get_enum_attribute(0, 1).unwrap(), enum_attribute);

pub fn get_string_attribute(&self, index: u32, key: &str) -> Option<Attribute>

Gets a string Attribute on this CallSiteValue at an index and key.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.add_attribute(0, string_attribute);
call_site_value.add_attribute(0, enum_attribute);

assert_eq!(call_site_value.get_string_attribute(0, "my_key").unwrap(), string_attribute);

pub fn remove_enum_attribute(&self, index: u32, kind_id: u32)

Removes an enum Attribute on this CallSiteValue at an index and kind id.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.add_attribute(0, string_attribute);
call_site_value.add_attribute(0, enum_attribute);
call_site_value.remove_enum_attribute(0, 1);

assert_eq!(call_site_value.get_enum_attribute(0, 1), None);

pub fn remove_string_attribute(&self, index: u32, key: &str)

Removes a string Attribute on this CallSiteValue at an index and key.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.add_attribute(0, string_attribute);
call_site_value.add_attribute(0, enum_attribute);
call_site_value.remove_string_attribute(0, "my_key");

assert_eq!(call_site_value.get_string_attribute(0, "my_key"), None);

pub fn count_arguments(&self) -> u32

Counts the number of arguments this CallSiteValue was called with.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

assert_eq!(call_site_value.count_arguments(), 0);

pub fn get_call_convention(&self) -> u32

Gets the calling convention for this CallSiteValue.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

assert_eq!(call_site_value.get_call_convention(), 0);

pub fn set_call_convention(&self, conv: u32)

Sets the calling convention for this CallSiteValue.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.set_call_convention(2);

assert_eq!(call_site_value.get_call_convention(), 2);

pub fn set_param_alignment_attribute(&self, index: u32, alignment: u32)

Shortcut for setting the alignment Attribute for this CallSiteValue.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = fn_value.append_basic_block("entry");

builder.position_at_end(&entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn");

call_site_value.set_param_alignment_attribute(0, 2);

pub fn print_to_string(&self) -> LLVMString

Prints the definition of a CallSiteValue to a LLVMString.