Struct inkwell::values::FloatValue

pub struct FloatValue<'ctx> { /* fields omitted */ }

Implementations

impl<'ctx> FloatValue<'ctx>

pub fn get_name(&self) -> &CStr

Gets the name of a FloatValue. If the value is a constant, this will return an empty string.

pub fn get_type(self) -> FloatType<'ctx>

pub fn is_null(self) -> bool

pub fn is_undef(self) -> bool

pub fn print_to_stderr(self)

pub fn as_instruction(self) -> Option<InstructionValue<'ctx>>

pub fn const_neg(self) -> Self

pub fn const_add(self, rhs: FloatValue<'ctx>) -> Self

pub fn const_sub(self, rhs: FloatValue<'ctx>) -> Self

pub fn const_mul(self, rhs: FloatValue<'ctx>) -> Self

pub fn const_div(self, rhs: FloatValue<'ctx>) -> Self

pub fn const_remainder(self, rhs: FloatValue<'ctx>) -> Self

pub fn const_cast(self, float_type: FloatType<'ctx>) -> Self

pub fn const_to_unsigned_int(self, int_type: IntType<'ctx>) -> IntValue<'ctx>

pub fn const_to_signed_int(self, int_type: IntType<'ctx>) -> IntValue<'ctx>

pub fn const_truncate(self, float_type: FloatType<'ctx>) -> FloatValue<'ctx>

pub fn const_extend(self, float_type: FloatType<'ctx>) -> FloatValue<'ctx>

pub fn const_compare(
self,
op: FloatPredicate,
rhs: FloatValue<'ctx>
) -> IntValue<'ctx>

pub fn is_const(self) -> bool

Determines whether or not a FloatValue is a constant.

Example

use inkwell::context::Context;

let context = Context::create();
let f64_type = context.f64_type();
let f64_val = f64_type.const_float(1.2);

assert!(f64_val.is_const());

pub fn get_constant(self) -> Option<(f64, bool)>

Obtains a constant FloatValue's value and whether or not it lost info.

Example

use inkwell::context::Context;

let context = Context::create();
let f64_type = context.f64_type();
let f64_1_2 = f64_type.const_float(1.2);

assert_eq!(f64_1_2.get_constant(), Some((1.2, false)));

pub fn replace_all_uses_with(self, other: FloatValue<'ctx>)