Struct inkwell::values::FunctionValue

pub struct FunctionValue { /* fields omitted */ }

Methods

impl FunctionValue

pub fn get_linkage(&self) -> Linkage

pub fn set_linkage(&self, linkage: Linkage)

pub fn is_null(&self) -> bool

pub fn is_undef(&self) -> bool

pub fn print_to_string(&self) -> LLVMString

pub fn print_to_stderr(&self)

pub fn verify(&self, print: bool) -> bool

pub fn get_next_function(&self) -> Option<Self>

pub fn get_previous_function(&self) -> Option<Self>

pub fn get_first_param(&self) -> Option<BasicValueEnum>

pub fn get_last_param(&self) -> Option<BasicValueEnum>

pub fn get_entry_basic_block(&self) -> Option<BasicBlock>

pub fn get_first_basic_block(&self) -> Option<BasicBlock>

pub fn append_basic_block(&self, name: &str) -> BasicBlock

pub fn get_nth_param(&self, nth: u32) -> Option<BasicValueEnum>

pub fn count_params(&self) -> u32

pub fn count_basic_blocks(&self) -> u32

pub fn get_basic_blocks(&self) -> Vec<BasicBlock>

pub fn get_return_type(&self) -> BasicTypeEnum

pub fn get_param_iter(&self) -> ParamValueIter

pub fn get_params(&self) -> Vec<BasicValueEnum>

pub fn get_last_basic_block(&self) -> Option<BasicBlock>

pub fn get_name(&self) -> &CStr

pub fn view_function_config(&self)

pub fn view_function_config_only(&self)

pub unsafe fn delete(self)

pub fn get_type(&self) -> FunctionType

pub fn has_metadata(&self) -> bool

pub fn get_metadata(&self, kind_id: u32) -> Option<MetadataValue>

pub fn set_metadata(&self, metadata: MetadataValue, kind_id: u32)

pub fn has_personality_function(&self) -> bool

pub fn get_personality_function(&self) -> Option<FunctionValue>

pub fn set_personality_function(&self, personality_fn: FunctionValue)

pub fn get_intrinsic_id(&self) -> u32

pub fn get_call_conventions(&self) -> u32

pub fn set_call_conventions(&self, call_conventions: u32)

pub fn get_gc(&self) -> &CStr

pub fn set_gc(&self, gc: &str)

pub fn replace_all_uses_with(&self, other: FunctionValue)

pub fn add_attribute(&self, index: u32, attribute: Attribute)

Adds an Attribute to this FunctionValue if index is 0, otherwise to its parameter in the 1 - Nth position.

Example

use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(0, string_attribute);
fn_value.add_attribute(0, enum_attribute);

pub fn count_attributes(&self, index: u32) -> u32

Counts the number of Attributes belonging to this FunctionValue if index is 0, otherwise to its parameter in the 1 - Nth position.

Example

use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(0, string_attribute);
fn_value.add_attribute(0, enum_attribute);

assert_eq!(fn_value.count_attributes(0), 2);

pub fn remove_string_attribute(&self, index: u32, key: &str)

Removes a string Attribute belonging this FunctionValue if index is 0, otherwise to its parameter in the 1 - Nth position.

Example

use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");

fn_value.add_attribute(0, string_attribute);
fn_value.remove_string_attribute(0, "my_key");

pub fn remove_enum_attribute(&self, index: u32, kind_id: u32)

Removes an enum Attribute belonging to this FunctionValue if index is 0, otherwise to its parameter in the 1 - Nth position.

Example

use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(0, enum_attribute);
fn_value.remove_enum_attribute(0, 1);

pub fn get_enum_attribute(&self, index: u32, kind_id: u32) -> Option<Attribute>

Gets an enum Attribute belonging to this FunctionValue if index is 0, otherwise to its parameter in the 1 - Nth position.

Example

use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(0, enum_attribute);

assert_eq!(fn_value.get_enum_attribute(0, 1), Some(enum_attribute));

pub fn get_string_attribute(&self, index: u32, key: &str) -> Option<Attribute>

Gets a string Attribute belonging this FunctionValue if index is 0, otherwise to its parameter in the 1 - Nth position.

Example

use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");

fn_value.add_attribute(0, string_attribute);

assert_eq!(fn_value.get_string_attribute(0, "my_key"), Some(string_attribute));

pub fn set_param_alignment(&self, param_index: u32, alignment: u32)

pub fn as_global_value(&self) -> GlobalValue

Gets the GlobalValue version of this FunctionValue. This allows you to further inspect its global properties or even convert it to a PointerValue.