Struct inkwell::values::PhiValue

pub struct PhiValue<'ctx> { /* fields omitted */ }

A Phi Instruction returns a value based on which basic block branched into the Phi's containing basic block.

Implementations

impl<'ctx> PhiValue<'ctx>

pub fn add_incoming(
self,
incoming: &[(&dyn BasicValue<'ctx>, BasicBlock<'ctx>)]
)

pub fn count_incoming(self) -> u32

pub fn get_incoming(
self,
index: u32
) -> Option<(BasicValueEnum<'ctx>, BasicBlock<'ctx>)>

pub fn get_name(&self) -> &CStr

Gets the name of a ArrayValue. If the value is a constant, this will return an empty string.

pub fn set_name(self, name: &str)

pub fn is_null(self) -> bool

pub fn is_undef(self) -> bool

pub fn as_instruction(self) -> InstructionValue<'ctx>

pub fn replace_all_uses_with(self, other: &PhiValue<'ctx>)

pub fn as_basic_value(self) -> BasicValueEnum<'ctx>