inkwell::values

Struct PointerValue

pub struct PointerValue<'ctx> { /* private fields */ }

Implementations

impl<'ctx> PointerValue<'ctx>

pub unsafe fn new(value: LLVMValueRef) -> Self

Get a value from an LLVMValueRef.

Safety

The ref must be valid and of type pointer.

pub fn get_name(&self) -> &CStr

Get name of the PointerValue. If the value is a constant, this will return an empty string.

pub fn set_name(&self, name: &str)

Set name of the PointerValue.

pub fn get_type(self) -> PointerType<'ctx>

pub fn is_null(self) -> bool

pub fn is_undef(self) -> bool

pub fn is_const(self) -> bool

Determines whether or not a PointerValue is a constant.

Example

use inkwell::AddressSpace;
use inkwell::context::Context;

let context = Context::create();
let void_type = context.void_type();

pub fn print_to_stderr(self)

pub fn as_instruction(self) -> Option<InstructionValue<'ctx>>

pub unsafe fn const_gep<T: BasicType<'ctx>>( self, ty: T, ordered_indexes: &[IntValue<'ctx>], ) -> PointerValue<'ctx>

Available on non-crate feature typed-pointers only.

GEP is very likely to segfault if indexes are used incorrectly, and is therefore an unsafe function. Maybe we can change this in the future.

pub unsafe fn const_in_bounds_gep<T: BasicType<'ctx>>( self, ty: T, ordered_indexes: &[IntValue<'ctx>], ) -> PointerValue<'ctx>

Available on non-crate feature typed-pointers only.

GEP is very likely to segfault if indexes are used incorrectly, and is therefore an unsafe function. Maybe we can change this in the future.

impl<'ctx> PointerValue<'ctx>

pub fn const_to_int(self, int_type: IntType<'ctx>) -> IntValue<'ctx>

pub fn const_cast(self, ptr_type: PointerType<'ctx>) -> PointerValue<'ctx>

pub fn const_address_space_cast( self, ptr_type: PointerType<'ctx>, ) -> PointerValue<'ctx>

pub fn replace_all_uses_with(self, other: PointerValue<'ctx>)