inkwell::values

Struct StructValue

pub struct StructValue<'ctx> { /* private fields */ }

Implementations

impl<'ctx> StructValue<'ctx>

pub unsafe fn new(value: LLVMValueRef) -> Self

Get a value from an LLVMValueRef.

Safety

The ref must be valid and of type struct.

pub fn get_field_at_index(self, index: u32) -> Option<BasicValueEnum<'ctx>>

Gets the value of a field belonging to this StructValue.

use inkwell::context::Context;

let context = Context::create();
let i32_type = context.i32_type();
let i8_type = context.i8_type();
let i8_val = i8_type.const_all_ones();
let i32_val = i32_type.const_all_ones();
let struct_type = context.struct_type(&[i8_type.into(), i32_type.into()], false);
let struct_val = struct_type.const_named_struct(&[i8_val.into(), i32_val.into()]);

assert!(struct_val.get_field_at_index(0).is_some());
assert!(struct_val.get_field_at_index(1).is_some());
assert!(struct_val.get_field_at_index(3).is_none());
assert!(struct_val.get_field_at_index(0).unwrap().is_int_value());

pub unsafe fn get_field_at_index_unchecked( self, index: u32, ) -> BasicValueEnum<'ctx>

Gets the value of a field belonging to this StructValue.

Safety

The index must be smaller than StructValue::count_fields.

pub fn get_fields(self) -> FieldValueIter<'ctx>

Get a field value iterator.

pub fn set_field_at_index<BV: BasicValue<'ctx>>( self, index: u32, val: BV, ) -> bool

Sets the value of a field belonging to this StructValue.

pub fn count_fields(self) -> u32

Counts the number of fields.

use inkwell::context::Context;

let context = Context::create();
let i32_type = context.i32_type();
let i8_type = context.i8_type();
let i8_val = i8_type.const_all_ones();
let i32_val = i32_type.const_all_ones();
let struct_type = context.struct_type(&[i8_type.into(), i32_type.into()], false);
let struct_val = struct_type.const_named_struct(&[i8_val.into(), i32_val.into()]);

assert_eq!(struct_val.count_fields(), 2);
assert_eq!(struct_val.count_fields(), struct_type.count_fields());

pub fn get_name(&self) -> &CStr

Gets the name of a StructValue. If the value is a constant, this will return an empty string.

pub fn set_name(&self, name: &str)

Get name of the StructValue.

pub fn get_type(self) -> StructType<'ctx>

pub fn is_null(self) -> bool

pub fn is_undef(self) -> bool

pub fn print_to_stderr(self)

pub fn as_instruction(self) -> Option<InstructionValue<'ctx>>

pub fn replace_all_uses_with(self, other: StructValue<'ctx>)

pub fn is_const(self) -> bool

Determines whether or not a StructValue is a constant.

Example

use inkwell::{context::Context, values::BasicValue};

let context = Context::create();
let i64_type = context.i64_type();
let i64_val = i64_type.const_int(23, false).as_basic_value_enum();
let struct_val = context.const_struct(&[i64_val, i64_val], false);

assert!(struct_val.is_const());