Struct inkwell::values::VectorValue

pub struct VectorValue<'ctx> { /* fields omitted */ }

Implementations

impl<'ctx> VectorValue<'ctx>

pub fn is_const(self) -> bool

Determines whether or not a VectorValue is a constant.

Example

use inkwell::context::Context;

let context = Context::create();
let i8_type = context.i8_type();
let i8_vec_type = i8_type.vec_type(3);
let i8_vec_zero = i8_vec_type.const_zero();

assert!(i8_vec_zero.is_const());

pub fn is_constant_vector(self) -> bool

pub fn is_constant_data_vector(self) -> bool

pub fn print_to_stderr(self)

pub fn get_name(&self) -> &CStr

Gets the name of a VectorValue. If the value is a constant, this will return an empty string.

pub fn get_type(self) -> VectorType<'ctx>

pub fn is_null(self) -> bool

pub fn is_undef(self) -> bool

pub fn as_instruction(self) -> Option<InstructionValue<'ctx>>

pub fn const_extract_element(
self,
index: IntValue<'ctx>
) -> BasicValueEnum<'ctx>

pub fn const_insert_element<BV: BasicValue<'ctx>>(
self,
index: IntValue<'ctx>,
value: BV
) -> BasicValueEnum<'ctx>

pub fn replace_all_uses_with(self, other: VectorValue<'ctx>)

pub fn is_const_string(self) -> bool

Creates a const string which may be null terminated.

Example

use inkwell::context::Context;

let context = Context::create();
let string = context.const_string(b"my_string", false);

assert!(string.is_const_string());

pub fn get_string_constant(&self) -> &CStr

pub fn get_element_as_constant(self, index: u32) -> BasicValueEnum<'ctx>

pub fn const_select<BV: BasicValue<'ctx>>(
self,
then: BV,
else_: BV
) -> BasicValueEnum<'ctx>

pub fn const_shuffle_vector(
self,
right: VectorValue<'ctx>,
mask: VectorValue<'ctx>
) -> VectorValue<'ctx>