Trait inkwell::values::AggregateValue

pub unsafe trait AggregateValue<'ctx>: BasicValue<'ctx> {
    // Provided method
    fn as_aggregate_value_enum(&self) -> AggregateValueEnum<'ctx> { ... }
}

Represents an aggregate value, built on top of other values.

Provided Methods

fn as_aggregate_value_enum(&self) -> AggregateValueEnum<'ctx>

Returns an enum containing a typed version of the AggregateValue.