Crate inkwell

Inkwell documentation is a work in progress.

If you have any LLVM knowledge that could be used to improve these docs, we would greatly appreciate you opening an issue and/or a pull request on our GitHub page.

Due to a rustdoc issue, this documentation represents only the latest supported LLVM version. We hope that this issue will be resolved in the future so that multiple versions can be documented side by side.

Modules

basic_block
A BasicBlock is a container of instructions.

builder

context

data_layout

execution_engine

memory_buffer

module

object_file

passes

targets

types

values

Enums

AddressSpace
Defines the address space in which a global will be inserted.

DLLStorageClass

FloatPredicate
Defines how to compare a left and right float value.

GlobalVisibility

IntPredicate

OptimizationLevel
Defines the optimization level used to compile a Module.

ThreadLocalMode

Functions

enable_llvm_pretty_stack_trace

is_multithreaded

load_library_permanently

reset_fatal_error_handler
Resets LLVM's fatal error handler back to the default

shutdown_llvm