inkwell::memory_buffer

Struct MemoryBuffer

pub struct MemoryBuffer<'a> { /* private fields */ }

Implementations

impl MemoryBuffer<'static>

pub fn create_from_file(path: &Path) -> Result<Self, LLVMString>

Create a memory buffer from file content.

Return Err with error message if failed.

pub fn create_from_stdin() -> Result<Self, LLVMString>

Create a memory buffer from stdin.

Return Err with error message if failed.

pub fn create_from_memory_range_copy(input: &[u8], name: &str) -> Self

Create a memory buffer copied from a byte slice with a trailing nul byte.

Panics

Panics if the input byte slice does not terminate with a nul byte.

impl<'a> MemoryBuffer<'a>

pub unsafe fn new(memory_buffer: LLVMMemoryBufferRef) -> Self

pub fn as_mut_ptr(&self) -> LLVMMemoryBufferRef

pub fn create_from_memory_range(input: &'a [u8], name: &str) -> Self

Create a memory buffer from a byte slice with a trailing nul byte.

Panics

Panics if the input byte slice does not terminate with a nul byte.

pub fn as_slice(&self) -> &[u8]

Gets a byte slice of this MemoryBuffer, containing the trailing nul byte.

pub fn get_size(&self) -> usize

Gets the byte size of this MemoryBuffer, counting the trailing nul byte.

pub fn create_binary_file( &self, context: Option<&Context>, ) -> Result<BinaryFile<'_>, LLVMString>

Convert this MemoryBuffer and optional Context into a BinaryFile.

The context is required if the resulting file is LLVM IR.

Return Err with error message if failed.