Struct inkwell::module::Module

pub struct Module { /* fields omitted */ }

Represents a reference to an LLVM Module. The underlying module will be disposed when dropping this object.

Methods

impl Module

pub fn create(name: &str) -> Self

Creates a named Module. Will be automatically assigned the global context.

To use your own Context, see inkwell::context::create_module()

Example

use inkwell::context::Context;
use inkwell::module::Module;

let context = Context::get_global();
let module = Module::create("my_module");

assert_eq!(module.get_context(), context);

pub fn add_function(
&self,
name: &str,
ty: &FunctionType,
linkage: Option<&Linkage>
) -> FunctionValue

Creates a function given its name and ty, adds it to the Module and returns it.

An optional linkage can be specified, without which the default value Linkage::ExternalLinkage will be used.

Example

use inkwell::context::Context;
use inkwell::module::{Module, Linkage};

let context = Context::get_global();
let module = Module::create("my_module");

let fn_type = context.f32_type().fn_type(&[], false);
let fn_val = module.add_function("my_function", &fn_type, None);

assert_eq!(fn_val.get_name().to_str(), Ok("my_function"));
assert_eq!(fn_val.get_linkage(), Linkage::ExternalLinkage);

pub fn get_context(&self) -> ContextRef

Gets the Context from which this Module originates.

Example

use inkwell::context::{Context, ContextRef};
use inkwell::module::Module;

let global_context = Context::get_global();
let global_module = Module::create("my_global_module");

assert_eq!(global_module.get_context(), global_context);

let local_context = Context::create();
let local_module = local_context.create_module("my_module");

assert_eq!(*local_module.get_context(), local_context);
assert_ne!(local_context, *global_context);

pub fn get_first_function(&self) -> Option<FunctionValue>

Gets the first FunctionValue defined in this Module.

Example

use inkwell::context::Context;
use inkwell::module::Module;

let context = Context::create();
let module = context.create_module("my_mod");

assert!(module.get_first_function().is_none());

let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", &fn_type, None);

assert_eq!(fn_value, module.get_first_function().unwrap());

pub fn get_last_function(&self) -> Option<FunctionValue>

Gets the last FunctionValue defined in this Module.

Example

use inkwell::context::Context;
use inkwell::module::Module;

let context = Context::create();
let module = context.create_module("my_mod");

assert!(module.get_last_function().is_none());

let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", &fn_type, None);

assert_eq!(fn_value, module.get_last_function().unwrap());

pub fn get_function(&self, name: &str) -> Option<FunctionValue>

Gets a FunctionValue defined in this Module by its name.

Example

use inkwell::context::Context;
use inkwell::module::Module;

let context = Context::create();
let module = context.create_module("my_mod");

assert!(module.get_function("my_fn").is_none());

let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", &fn_type, None);

assert_eq!(fn_value, module.get_function("my_fn").unwrap());

pub fn get_type(&self, name: &str) -> Option<BasicTypeEnum>

pub fn set_target(&self, target_triple: &str)

pub fn get_target(&self) -> &CStr

pub fn create_jit_execution_engine(
&self,
opt_level: OptimizationLevel
) -> Result<ExecutionEngine, String>

Creates a JIT ExecutionEngine from this Module.

Example

use inkwell::OptimizationLevel;
use inkwell::context::Context;
use inkwell::module::Module;
use inkwell::targets::{InitializationConfig, Target};

Target::initialize_native(&InitializationConfig::default()).expect("Failed to initialize native target");

let context = Context::get_global();
let module = Module::create("my_module");
let execution_engine = module.create_jit_execution_engine(OptimizationLevel::None).unwrap();

assert_eq!(module.get_context(), context);

pub fn add_global(
&self,
type_: &BasicType,
address_space: Option<AddressSpace>,
name: &str
) -> GlobalValue

pub fn write_bitcode_to_path(&self, path: &Path) -> bool

pub fn write_bitcode_to_file(
&self,
file: &File,
should_close: bool,
unbuffered: bool
) -> bool

pub fn write_bitcode_to_memory(&self) -> MemoryBuffer

pub fn verify(&self, print: bool) -> bool

Ensures that the current Module is valid, and returns a bool that describes whether or not it is.

print - Whether or not a message describing the error should be printed to stderr, in case of failure.

Remarks

See also: http://llvm.org/doxygen/Analysis_2Analysis_8cpp_source.html

pub fn get_data_layout(&self) -> &DataLayout

pub fn set_data_layout(&self, data_layout: &DataLayout)

pub fn print_to_stderr(&self)

Prints the content of the Module to stderr.

pub fn print_to_string(&self) -> &CStr

Prints the content of the Module to a string.

pub fn print_to_file(&self, path: &Path) -> Result<(), String>

Prints the content of the Module to a file.

pub fn set_inline_assembly(&self, asm: &str)

pub fn add_global_metadata(&self, key: &str, metadata: &MetadataValue)

pub fn get_global_metadata_size(&self, key: &str) -> u32

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn get_global_metadata(&self, key: &str) -> Vec<MetadataValue>

pub fn get_first_global(&self) -> Option<GlobalValue>

pub fn get_last_global(&self) -> Option<GlobalValue>

pub fn get_global(&self, name: &str) -> Option<GlobalValue>