Module inkwell::support

Modules

error_handling

This module contains some supplemental functions for dealing with errors.

Structs

LLVMString

An owned LLVM String. Also known as a LLVM Message

Functions

enable_llvm_pretty_stack_trace

is_multithreaded

Determines whether or not LLVM has been configured to run in multithreaded mode. (Inkwell currently does not officially support multithreaded mode)

load_library_permanently

shutdown_llvm⚠

This function is very unsafe. Any reference to LLVM data after this function is called will likey segfault. Probably only ever useful to call before your program ends. Might not even be absolutely necessary.