inkwell
Module support

Modules

error_handling

This module contains some supplemental functions for dealing with errors.

Structs

LLVMString

An owned LLVM String. Also known as a LLVM Message

Enums

LoadLibraryError

Possible errors that can occur when loading a library

Functions

enable_llvm_pretty_stack_trace

get_llvm_versionllvm16-0 or llvm17-0 or llvm18-0

Returns the major, minor, and patch version of the LLVM in use

is_multithreaded

Determines whether or not LLVM has been configured to run in multithreaded mode. (Inkwell currently does not officially support multithreaded mode)

load_library_permanently

Permanently load the dynamic library at the given path.

load_visible_symbols

Permanently loads all the symbols visible inside the current program

search_for_address_of_symbol

Search through all previously loaded dynamic libraries for symbol.

shutdown_llvm⚠

This function is very unsafe. Any reference to LLVM data after this function is called will likely segfault. Probably only ever useful to call before your program ends. Might not even be absolutely necessary.