inkwell::support
Struct LLVMString

pub struct LLVMString { /* private fields */ }

An owned LLVM String. Also known as a LLVM Message

Implementations

impl LLVMString

pub fn to_string(&self) -> String

This is a convenience method for creating a Rust String, however; it will reallocate. LLVMString should be used as much as possible to save memory since it is allocated by LLVM. It’s essentially a CString with a custom LLVM deallocator