Module inkwell::types

A type is a classification which determines how data is used.

Structs

ArrayType
An ArrayType is the type of contiguous constants or variables.

FloatType
A FloatType is the type of a floating point constant or variable.

FunctionType
A FunctionType is the type of a function variable.

IntType
An IntType is the type of an integer constant or variable.

PointerType
A PointerType is the type of a pointer constant or variable.

StructType
A StructType is the type of a heterogeneous container of types.

VectorType
A VectorType is the type of a multiple value SIMD constant or variable.

VoidType
A VoidType is a special type with no possible direct instances. It's only useful as a function return type.

Enums

AnyTypeEnum
A wrapper for any BasicType, VoidType, or FunctionType.

BasicTypeEnum
A wrapper for any BasicType.

StringRadix
How to interpret a string or digits used to construct an integer constant.

Traits

AnyType
Represents any LLVM type.

BasicType
Represents a basic LLVM type, that may be used in functions and struct definitions.

FloatMathType
Represents an LLVM type that can have floating point math operations applied to it.

IntMathType
Represents an LLVM type that can have integer math operations applied to it.

PointerMathType
Represents an LLVM type that can have pointer operations applied to it.