inkwell::types

Struct StructType

pub struct StructType<'ctx> { /* private fields */ }

A StructType is the type of a heterogeneous container of types.

Implementations

impl<'ctx> StructType<'ctx>

pub unsafe fn new(struct_type: LLVMTypeRef) -> Self

Create StructType from LLVMTypeRef

Safety

Undefined behavior, if referenced type isnโt struct type

pub fn get_field_type_at_index(self, index: u32) -> Option<BasicTypeEnum<'ctx>>

Gets the type of a field belonging to this StructType.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into()], false);

assert_eq!(struct_type.get_field_type_at_index(0).unwrap().into_float_type(), f32_type);

pub unsafe fn get_field_type_at_index_unchecked( self, index: u32, ) -> BasicTypeEnum<'ctx>

Gets the type of a field belonging to this StructType.

Safety

The index must be less than StructType::count_fields and the struct must not be opaque.

pub fn const_named_struct( self, values: &[BasicValueEnum<'ctx>], ) -> StructValue<'ctx>

Creates a StructValue based on this StructTypeโs definition.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let f32_zero = f32_type.const_float(0.);
let struct_type = context.struct_type(&[f32_type.into()], false);
let struct_val = struct_type.const_named_struct(&[f32_zero.into()]);

pub fn const_zero(self) -> StructValue<'ctx>

Creates a constant zero value of this StructType.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_zero = struct_type.const_zero();

pub fn size_of(self) -> Option<IntValue<'ctx>>

Gets the size of this StructType. Value may vary depending on the target architecture.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let f32_struct_type = context.struct_type(&[f32_type.into()], false);
let f32_struct_type_size = f32_struct_type.size_of();

pub fn get_context(self) -> ContextRef<'ctx>

Gets a reference to the Context this StructType was created in.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);

assert_eq!(struct_type.get_context(), context);

pub fn get_name(&self) -> Option<&CStr>

Gets this StructTypeโs name.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.opaque_struct_type("opaque_struct");

assert_eq!(struct_type.get_name().unwrap().to_str().unwrap(), "opaque_struct");

pub fn ptr_type(self, address_space: AddressSpace) -> PointerType<'ctx>

๐Deprecated: Starting from version 15.0, LLVM doesnโt differentiate between pointer types. Use Context::ptr_type instead.

Creates a PointerType with this StructType for its element type.

Example

use inkwell::AddressSpace;
use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_ptr_type = struct_type.ptr_type(AddressSpace::default());

#[cfg(feature = "typed-pointers")]
assert_eq!(struct_ptr_type.get_element_type().into_struct_type(), struct_type);

pub fn fn_type( self, param_types: &[BasicMetadataTypeEnum<'ctx>], is_var_args: bool, ) -> FunctionType<'ctx>

Creates a FunctionType with this StructType for its return type.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let fn_type = struct_type.fn_type(&[], false);

pub fn array_type(self, size: u32) -> ArrayType<'ctx>

Creates an ArrayType with this StructType for its element type.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_array_type = struct_type.array_type(3);

assert_eq!(struct_array_type.len(), 3);
assert_eq!(struct_array_type.get_element_type().into_struct_type(), struct_type);

pub fn is_packed(self) -> bool

Determines whether or not a StructType is packed.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);

assert!(struct_type.is_packed());

pub fn is_opaque(self) -> bool

Determines whether or not a StructType is opaque.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.opaque_struct_type("opaque_struct");

assert!(struct_type.is_opaque());

pub fn count_fields(self) -> u32

Counts the number of field types.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let i8_type = context.i8_type();
let struct_type = context.struct_type(&[f32_type.into(), i8_type.into()], false);

assert_eq!(struct_type.count_fields(), 2);

pub fn get_field_types(self) -> Vec<BasicTypeEnum<'ctx>>

Gets this StructTypeโs field types.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let i8_type = context.i8_type();
let struct_type = context.struct_type(&[f32_type.into(), i8_type.into()], false);

assert_eq!(struct_type.get_field_types(), &[f32_type.into(), i8_type.into()]);

pub fn get_field_types_iter(self) -> FieldTypesIter<'ctx>

Get a struct field iterator.

pub fn print_to_string(self) -> LLVMString

Print the definition of a StructType to LLVMString.

pub fn get_undef(self) -> StructValue<'ctx>

Creates an undefined instance of a StructType.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let i8_type = context.i8_type();
let struct_type = context.struct_type(&[f32_type.into(), i8_type.into()], false);
let struct_type_undef = struct_type.get_undef();

assert!(struct_type_undef.is_undef());

pub fn get_poison(self) -> StructValue<'ctx>

Available on crate features llvm12-0 or llvm13-0 or llvm14-0 or llvm15-0 or llvm16-0 or llvm17-0 or llvm18-1 or llvm19-1 or llvm20-1 or llvm21-1 only.

Creates a poison instance of a StructType.

Example

use inkwell::context::Context;
use inkwell::values::AnyValue;

let context = Context::create();
let f32_type = context.f32_type();
let i8_type = context.i8_type();
let struct_type = context.struct_type(&[f32_type.into(), i8_type.into()], false);
let struct_type_poison = struct_type.get_poison();

assert!(struct_type_poison.is_poison());

pub fn set_body(self, field_types: &[BasicTypeEnum<'ctx>], packed: bool) -> bool

Defines the body of a StructType.

If the struct is an opaque type, it will no longer be after this call.

Resetting the packed state of a non-opaque struct type may not work.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let opaque_struct_type = context.opaque_struct_type("opaque_struct");

opaque_struct_type.set_body(&[f32_type.into()], false);

assert!(!opaque_struct_type.is_opaque());

pub fn const_array(self, values: &[StructValue<'ctx>]) -> ArrayValue<'ctx>

Creates a constant ArrayValue.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_val = struct_type.const_named_struct(&[]);
let struct_array = struct_type.const_array(&[struct_val, struct_val]);

assert!(struct_array.is_const());