Struct inkwell::types::StructType

pub struct StructType { /* fields omitted */ }

A StructType is the type of a heterogeneous container of types.

Methods

impl StructType

pub fn get_field_type_at_index(&self, index: u32) -> Option<BasicTypeEnum>

Gets the type of a field belonging to this StructType.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into()], false);

assert_eq!(struct_type.get_field_type_at_index(0).unwrap().into_float_type(), f32_type);

pub fn const_named_struct(&self, values: &[BasicValueEnum]) -> StructValue

Creates a StructValue based on this StructType's definition.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let f32_zero = f32_type.const_float(0.);
let struct_type = context.struct_type(&[f32_type.into()], false);
let struct_val = struct_type.const_named_struct(&[f32_zero.into()]);

pub fn const_struct(values: &[BasicValueEnum], packed: bool) -> StructValue

Creates a StructValue based on the input values rather than an existing StructType. It will be assigned the global Context.

Example

use inkwell::types::{FloatType, StructType};

let f32_type = FloatType::f32_type();
let f32_zero = f32_type.const_float(0.);
let struct_val = StructType::const_struct(&[f32_zero.into()], false);

pub fn const_zero(&self) -> StructValue

Creates a constant zero value of this StructType.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_zero = struct_type.const_zero();

pub fn is_sized(&self) -> bool

Gets whether or not this StructType is sized or not.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let f32_struct_type = context.struct_type(&[f32_type.into()], false);

assert!(f32_struct_type.is_sized());

pub fn size_of(&self) -> Option<IntValue>

Gets the size of this StructType. Value may vary depending on the target architecture.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let f32_struct_type = context.struct_type(&[f32_type.into()], false);
let f32_struct_type_size = f32_struct_type.size_of();

pub fn get_alignment(&self) -> IntValue

Gets the alignment of this StructType. Value may vary depending on the target architecture.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_type_alignment = struct_type.get_alignment();

pub fn get_context(&self) -> ContextRef

Gets a reference to the Context this StructType was created in.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);

assert_eq!(*struct_type.get_context(), context);

pub fn get_name(&self) -> Option<&CStr>

Gets this StructType's name.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.opaque_struct_type("opaque_struct");

assert_eq!(struct_type.get_name().unwrap().to_str().unwrap(), "opaque_struct");

pub fn ptr_type(&self, address_space: AddressSpace) -> PointerType

Creates a PointerType with this StructType for its element type.

Example

use inkwell::AddressSpace;
use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_ptr_type = struct_type.ptr_type(AddressSpace::Generic);

assert_eq!(struct_ptr_type.get_element_type().into_struct_type(), struct_type);

pub fn fn_type(
&self,
param_types: &[BasicTypeEnum],
is_var_args: bool
) -> FunctionType

Creates a FunctionType with this StructType for its return type.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let fn_type = struct_type.fn_type(&[], false);

pub fn array_type(&self, size: u32) -> ArrayType

Creates an ArrayType with this StructType for its element type.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_array_type = struct_type.array_type(3);

assert_eq!(struct_array_type.len(), 3);
assert_eq!(struct_array_type.get_element_type().into_struct_type(), struct_type);

pub fn is_packed(&self) -> bool

Determines whether or not a StructType is packed.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);

assert!(struct_type.is_packed());

pub fn is_opaque(&self) -> bool

Determines whether or not a StructType is opaque.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.opaque_struct_type("opaque_struct");

assert!(struct_type.is_opaque());

pub fn struct_type(field_types: &[BasicTypeEnum], packed: bool) -> Self

Creates a StructType definiton from heterogeneous types in the global Context.

Example

use inkwell::context::Context;
use inkwell::types::{FloatType, IntType, StructType};

let f32_type = FloatType::f32_type();
let i8_type = IntType::i8_type();
let struct_type = StructType::struct_type(&[f32_type.into(), i8_type.into()], false);

assert_eq!(struct_type.get_context(), Context::get_global());

pub fn count_fields(&self) -> u32

Counts the number of field types.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let i8_type = context.i8_type();
let struct_type = context.struct_type(&[f32_type.into(), i8_type.into()], false);

assert_eq!(struct_type.count_fields(), 2);

pub fn get_field_types(&self) -> Vec<BasicTypeEnum>

Gets this StructType's field types.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let i8_type = context.i8_type();
let struct_type = context.struct_type(&[f32_type.into(), i8_type.into()], false);

assert_eq!(struct_type.get_field_types(), &[f32_type.into(), i8_type.into()]);

pub fn print_to_string(&self) -> LLVMString

Prints the definition of a StructType to a LLVMString.

pub fn get_undef(&self) -> StructValue

Creates an undefined instance of a StructType.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let i8_type = context.i8_type();
let struct_type = context.struct_type(&[f32_type.into(), i8_type.into()], false);
let struct_type_undef = struct_type.get_undef();

assert!(struct_type_undef.is_undef());

pub fn set_body(&self, field_types: &[BasicTypeEnum], packed: bool) -> bool

Determines whether or not a StructType is opaque.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let opaque_struct_type = context.opaque_struct_type("opaque_struct");

opaque_struct_type.set_body(&[f32_type.into()], false);

assert!(!opaque_struct_type.is_opaque());

pub fn const_array(&self, values: &[StructValue]) -> ArrayValue

Creates a constant ArrayValue.

Example

use inkwell::context::Context;

let context = Context::create();
let f32_type = context.f32_type();
let struct_type = context.struct_type(&[f32_type.into(), f32_type.into()], false);
let struct_val = struct_type.const_named_struct(&[]);
let struct_array = struct_type.const_array(&[struct_val, struct_val]);

assert!(struct_array.is_const());