Trait inkwell::types::IntMathType

pub trait IntMathType<'ctx>: BasicType<'ctx> {
    type ValueType: IntMathValue<'ctx>;
    type MathConvType: FloatMathType<'ctx>;
    type PtrConvType: PointerMathType<'ctx>;
}

Represents an LLVM type that can have integer math operations applied to it.

Associated Types

type ValueType: IntMathValue<'ctx>

The value instance of an int or int vector type.

type MathConvType: FloatMathType<'ctx>

The type for int to float or int vector to float vector conversions.

type PtrConvType: PointerMathType<'ctx>

The type for int to pointer or int vector to pointer vector conversions.