Trait inkwell::types::PointerMathType

pub trait PointerMathType<'ctx>: BasicType<'ctx> {
    type ValueType: PointerMathValue<'ctx>;
    type PtrConvType: IntMathType<'ctx>;
}

Represents an LLVM type that can have pointer operations applied to it.

Associated Types

type ValueType: PointerMathValue<'ctx>

The value instance of a pointer type.

type PtrConvType: IntMathType<'ctx>

The type for pointer to int or pointer vector to int conversions.

Loading content...