Module inkwell::values

A value is an instance of a type.

Structs

ArrayValue

An ArrayValue is a block of contiguous constants or variables.

BasicValueUse

A usage of a BasicValue in another value.

CallSiteValue

A value resulting from a function call. It may have function attributes applied to it.

FloatValue

FunctionValue

GenericValue

GlobalValue

InstructionValue

IntValue

MetadataValue

PhiValue

A Phi Instruction returns a value based on which basic block branched into the Phi’s containing basic block.

PointerValue

StructValue

VectorValue

Enums

AggregateValueEnum

AnyValueEnum

BasicMetadataValueEnum

BasicValueEnum

InstructionOpcode

UnnamedAddressllvm7-0 or llvm8-0 or llvm9-0 or llvm10-0 or llvm11-0 or llvm12-0 or llvm13-0 or llvm14-0 or llvm15-0

This enum determines the significance of a GlobalValue’s address.

Constants

FIRST_CUSTOM_METADATA_KIND_ID

Traits

AggregateValue

Represents an aggregate value, built on top of other values.

AnyValue

Defines any struct wrapping an LLVM value.

AsValueRef

BasicValue

Represents a basic value, which can be used both by itself, or in an AggregateValue.

FloatMathValue

Represents a value which is permitted in floating point math operations

IntMathValue

Represents a value which is permitted in integer math operations

PointerMathValue