Module inkwell::values

Structs

ArrayValue

FloatValue

FunctionValue

GenericValue

GlobalValue

InstructionValue

IntValue

MetadataValue

PhiValue

PointerValue

StructValue

VectorValue

Enums

AggregateValueEnum

AnyValueEnum

BasicMetadataValueEnum

BasicValueEnum

InstructionOpcode

Constants

FIRST_CUSTOM_METADATA_KIND_ID

Traits

AggregateValue
Represents an aggregate value, built on top of other values.

AnyValue
Defines any struct wrapping an LLVM value.

BasicValue
Represents a basic value, which can be used both by itself, or in an AggregateValue.