inkwell

Module values

A value is an instance of a type.

Structs

ArrayValue

An ArrayValue is a block of contiguous constants or variables.

BasicValueUse

A usage of a BasicValue in another value.

CallSiteValue

A value resulting from a function call. It may have function attributes applied to it.

FastMathFlagsllvm18-1 or llvm19-1 or llvm20-1 or llvm21-1

Fast math flags to enable otherwise unsafe floating-point transformations.

FieldValueIter

Iterate over all the field values of this struct.

FloatValue

FunctionValue

GenericValue

GlobalValue

IncomingIter

Iterate over all the incoming edges of a phi value.

InstructionValue

IntValue

MetadataValue

OperandBundlellvm18-1 or llvm19-1 or llvm20-1 or llvm21-1

One of an instruction’s operand bundles.

OperandIter

Iterate over all the operands of an instruction value.

OperandUseIter

Iterate over all the operands of an instruction value.

PhiValue

A Phi Instruction returns a value based on which basic block branched into the Phi’s containing basic block.

PointerValue

ScalableVectorValue

StructValue

VectorValue

Enums

AggregateValueEnum

AnyValueEnum

AtomicError

Errors for atomic operations on load/store instructions.

BasicMetadataValueEnum

BasicValueEnum

InstructionOpcode

InstructionValueError

Errors for InstructionValue.

LLVMTailCallKindllvm18-1 or llvm19-1 or llvm20-1 or llvm21-1

Tail call kind for LLVMSetTailCallKind and LLVMGetTailCallKind.

Operand

Either BasicValueEnum or BasicBlock.

UnnamedAddress

This enum determines the significance of a GlobalValue’s address.

ValueKind

Either BasicValueEnum or InstructionValue.

Constants

FIRST_CUSTOM_METADATA_KIND_ID

Value returned by Context::get_kind_id() for the first input string that isn’t known.

Traits

AggregateValue

Represents an aggregate value, built on top of other values.

AnyValue

Defines any struct wrapping an LLVM value.

AsValueRef

BasicValue

Represents a basic value, which can be used both by itself, or in an AggregateValue.

FloatMathValue

Represents a value which is permitted in floating point math operations

IntMathValue

Represents a value which is permitted in integer math operations

PointerMathValue

VectorBaseValue

Represents a value which is permitted in vector operations, either fixed or scalable