inkwell::values

Struct CallSiteValue

pub struct CallSiteValue<'ctx>(/* private fields */);

A value resulting from a function call. It may have function attributes applied to it.

This struct may be removed in the future in favor of an InstructionValue<CallSite> type.

Implementations

impl<'ctx> CallSiteValue<'ctx>

pub unsafe fn new(value: LLVMValueRef) -> Self

Get a value from an LLVMValueRef.

Safety

The ref must be valid and of type call site.

pub fn set_tail_call(self, tail_call: bool)

Sets whether or not this call is a tail call.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.set_tail_call(true);

pub fn is_tail_call(self) -> bool

Determines whether or not this call is a tail call.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.set_tail_call(true);

assert!(call_site_value.is_tail_call());

pub fn get_tail_call_kind(self) -> LLVMTailCallKind

Available on crate features llvm18-1 or llvm19-1 or llvm20-1 or llvm21-1 only.

Returns tail, musttail, and notail attributes.

Example

use inkwell::values::LLVMTailCallKind::*;

let context = inkwell::context::Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site = builder.build_call(fn_value, &[], "my_fn").unwrap();

assert_eq!(call_site.get_tail_call_kind(), LLVMTailCallKindNone);

pub fn set_tail_call_kind(self, kind: LLVMTailCallKind)

Available on crate features llvm18-1 or llvm19-1 or llvm20-1 or llvm21-1 only.

Sets tail, musttail, and notail attributes.

Example

use inkwell::values::LLVMTailCallKind::*;

let context = inkwell::context::Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site.set_tail_call_kind(LLVMTailCallKindTail);
assert_eq!(call_site.get_tail_call_kind(), LLVMTailCallKindTail);

pub fn try_as_basic_value(self) -> ValueKind<'ctx>

Try to convert this CallSiteValue to a BasicValueEnum if not a void return type.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

assert!(call_site_value.try_as_basic_value().is_instruction());

pub fn add_attribute(self, loc: AttributeLoc, attribute: Attribute)

Adds an Attribute to this CallSiteValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.add_attribute(AttributeLoc::Return, string_attribute);
call_site_value.add_attribute(AttributeLoc::Return, enum_attribute);

pub fn get_called_fn_value(self) -> Option<FunctionValue<'ctx>>

Gets the FunctionValue this CallSiteValue is based on.

Returns None if the call this value bases on is indirect or the retrieved function value doesn’t have the same type as the underlying call instruction.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

assert_eq!(call_site_value.get_called_fn_value(), Some(fn_value));

pub fn get_called_fn_type(self) -> FunctionType<'ctx>

Gets the type of the function called by the instruction this CallSiteValue is based on.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let i32_type = context.i32_type();
let fn_type = i32_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);

let entry_bb = context.append_basic_block(fn_value, "entry");
builder.position_at_end(entry_bb);

// Recursive call.
let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

assert_eq!(call_site_value.get_called_fn_type(), fn_type);

pub fn count_attributes(self, loc: AttributeLoc) -> u32

Counts the number of Attributes on this CallSiteValue at an index.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.add_attribute(AttributeLoc::Return, string_attribute);
call_site_value.add_attribute(AttributeLoc::Return, enum_attribute);

assert_eq!(call_site_value.count_attributes(AttributeLoc::Return), 2);

pub fn attributes(self, loc: AttributeLoc) -> Vec<Attribute>

Get all Attributes on this CallSiteValue at an index.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.add_attribute(AttributeLoc::Return, string_attribute);
call_site_value.add_attribute(AttributeLoc::Return, enum_attribute);

assert_eq!(call_site_value.attributes(AttributeLoc::Return), vec![ string_attribute, enum_attribute ]);

pub fn get_enum_attribute( self, loc: AttributeLoc, kind_id: u32, ) -> Option<Attribute>

Gets an enum Attribute on this CallSiteValue at an index and kind id.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.add_attribute(AttributeLoc::Return, string_attribute);
call_site_value.add_attribute(AttributeLoc::Return, enum_attribute);

assert_eq!(call_site_value.get_enum_attribute(AttributeLoc::Return, 1).unwrap(), enum_attribute);

pub fn get_string_attribute( self, loc: AttributeLoc, key: &str, ) -> Option<Attribute>

Gets a string Attribute on this CallSiteValue at an index and key.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.add_attribute(AttributeLoc::Return, string_attribute);
call_site_value.add_attribute(AttributeLoc::Return, enum_attribute);

assert_eq!(call_site_value.get_string_attribute(AttributeLoc::Return, "my_key").unwrap(), string_attribute);

pub fn remove_enum_attribute(self, loc: AttributeLoc, kind_id: u32)

Removes an enum Attribute on this CallSiteValue at an index and kind id.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.add_attribute(AttributeLoc::Return, string_attribute);
call_site_value.add_attribute(AttributeLoc::Return, enum_attribute);
call_site_value.remove_enum_attribute(AttributeLoc::Return, 1);

assert_eq!(call_site_value.get_enum_attribute(AttributeLoc::Return, 1), None);

pub fn remove_string_attribute(self, loc: AttributeLoc, key: &str)

Removes a string Attribute on this CallSiteValue at an index and key.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.add_attribute(AttributeLoc::Return, string_attribute);
call_site_value.add_attribute(AttributeLoc::Return, enum_attribute);
call_site_value.remove_string_attribute(AttributeLoc::Return, "my_key");

assert_eq!(call_site_value.get_string_attribute(AttributeLoc::Return, "my_key"), None);

pub fn count_arguments(self) -> u32

Counts the number of arguments this CallSiteValue was called with.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
// Enum attribute cannot have non-zero value
let enum_attribute = context.create_enum_attribute(1, 0);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

assert_eq!(call_site_value.count_arguments(), 0);

pub fn get_call_convention(self) -> u32

Gets the calling convention for this CallSiteValue.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

assert_eq!(call_site_value.get_call_convention(), 0);

pub fn set_call_convention(self, conv: u32)

Sets the calling convention for this CallSiteValue.

Example

use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.set_call_convention(2);

assert_eq!(call_site_value.get_call_convention(), 2);

pub fn set_alignment_attribute(self, loc: AttributeLoc, alignment: u32)

Shortcut for setting the alignment Attribute for this CallSiteValue.

Panics

When the alignment is not a power of 2.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let builder = context.create_builder();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let entry_bb = context.append_basic_block(fn_value, "entry");

builder.position_at_end(entry_bb);

let call_site_value = builder.build_call(fn_value, &[], "my_fn").unwrap();

call_site_value.set_alignment_attribute(AttributeLoc::Param(0), 2);

pub fn get_operand_bundles(&self) -> OperandBundleIter<'_, 'ctx>

Available on crate features llvm18-1 or llvm19-1 or llvm20-1 or llvm21-1 only.

Iterate over operand bundles.

Example

use inkwell::context::Context;
use inkwell::values::OperandBundle;

let context = Context::create();
let module = context.create_module("op_bundles");
let builder = context.create_builder();

let void_type = context.void_type();
let i32_type = context.i32_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("func", fn_type, None);

let basic_block = context.append_basic_block(fn_value, "entry");
builder.position_at_end(basic_block);

// Recursive call
let callinst = builder.build_direct_call_with_operand_bundles(
  fn_value,
  &[],
  &[OperandBundle::create("tag0", &[i32_type.const_zero().into()]), OperandBundle::create("tag1", &[])],
  "call"
).unwrap();

builder.build_return(None).unwrap();

let mut op_bundles_iter = callinst.get_operand_bundles();
assert_eq!(op_bundles_iter.len(), 2);
let tags: Vec<String> = op_bundles_iter.map(|ob| ob.get_tag().unwrap().into()).collect();
assert_eq!(tags, vec!["tag0", "tag1"]);