Struct inkwell::values::FunctionValue

pub struct FunctionValue<'ctx> { /* fields omitted */ }

Implementations

impl<'ctx> FunctionValue<'ctx>

pub fn get_linkage(self) -> Linkage

pub fn set_linkage(self, linkage: Linkage)

pub fn is_null(self) -> bool

pub fn is_undef(self) -> bool

pub fn print_to_stderr(self)

pub fn verify(self, print: bool) -> bool

pub fn get_next_function(self) -> Option<Self>

pub fn get_previous_function(self) -> Option<Self>

pub fn get_first_param(self) -> Option<BasicValueEnum<'ctx>>

pub fn get_last_param(self) -> Option<BasicValueEnum<'ctx>>

pub fn get_first_basic_block(self) -> Option<BasicBlock<'ctx>>

pub fn get_nth_param(self, nth: u32) -> Option<BasicValueEnum<'ctx>>

pub fn count_params(self) -> u32

pub fn count_basic_blocks(self) -> u32

pub fn get_basic_blocks(self) -> Vec<BasicBlock<'ctx>>

pub fn get_param_iter(self) -> ParamValueIter<'ctx>

pub fn get_params(self) -> Vec<BasicValueEnum<'ctx>>

pub fn get_last_basic_block(self) -> Option<BasicBlock<'ctx>>

pub fn get_name(&self) -> &CStr

Gets the name of a FunctionValue.

pub fn view_function_cfg(self)

View the control flow graph and produce a .dot file

pub fn view_function_cfg_only(self)

Only view the control flow graph

pub unsafe fn delete(self)

pub fn get_type(self) -> FunctionType<'ctx>

pub fn has_personality_function(self) -> bool

pub fn get_personality_function(self) -> Option<FunctionValue<'ctx>>

pub fn set_personality_function(self, personality_fn: FunctionValue<'ctx>)

pub fn get_intrinsic_id(self) -> u32

pub fn get_call_conventions(self) -> u32

pub fn set_call_conventions(self, call_conventions: u32)

pub fn get_gc(&self) -> &CStr

pub fn set_gc(self, gc: &str)

pub fn replace_all_uses_with(self, other: FunctionValue<'ctx>)

pub fn add_attribute(self, loc: AttributeLoc, attribute: Attribute)

Adds an Attribute to a particular location in this FunctionValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(AttributeLoc::Return, string_attribute);
fn_value.add_attribute(AttributeLoc::Return, enum_attribute);

pub fn count_attributes(self, loc: AttributeLoc) -> u32

Counts the number of Attributes belonging to the specified location in this FunctionValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(AttributeLoc::Return, string_attribute);
fn_value.add_attribute(AttributeLoc::Return, enum_attribute);

assert_eq!(fn_value.count_attributes(AttributeLoc::Return), 2);

pub fn attributes(self, loc: AttributeLoc) -> Vec<Attribute>

Get all Attributes belonging to the specified location in this FunctionValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(AttributeLoc::Return, string_attribute);
fn_value.add_attribute(AttributeLoc::Return, enum_attribute);

assert_eq!(fn_value.attributes(AttributeLoc::Return), vec![ string_attribute, enum_attribute ]);

pub fn remove_string_attribute(self, loc: AttributeLoc, key: &str)

Removes a string Attribute belonging to the specified location in this FunctionValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");

fn_value.add_attribute(AttributeLoc::Return, string_attribute);
fn_value.remove_string_attribute(AttributeLoc::Return, "my_key");

pub fn remove_enum_attribute(self, loc: AttributeLoc, kind_id: u32)

Removes an enum Attribute belonging to the specified location in this FunctionValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(AttributeLoc::Return, enum_attribute);
fn_value.remove_enum_attribute(AttributeLoc::Return, 1);

pub fn get_enum_attribute(
self,
loc: AttributeLoc,
kind_id: u32
) -> Option<Attribute>

Gets an enum Attribute belonging to the specified location in this FunctionValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let enum_attribute = context.create_enum_attribute(1, 1);

fn_value.add_attribute(AttributeLoc::Return, enum_attribute);

assert_eq!(fn_value.get_enum_attribute(AttributeLoc::Return, 1), Some(enum_attribute));

pub fn get_string_attribute(
self,
loc: AttributeLoc,
key: &str
) -> Option<Attribute>

Gets a string Attribute belonging to the specified location in this FunctionValue.

Example

use inkwell::attributes::AttributeLoc;
use inkwell::context::Context;

let context = Context::create();
let module = context.create_module("my_mod");
let void_type = context.void_type();
let fn_type = void_type.fn_type(&[], false);
let fn_value = module.add_function("my_fn", fn_type, None);
let string_attribute = context.create_string_attribute("my_key", "my_val");

fn_value.add_attribute(AttributeLoc::Return, string_attribute);

assert_eq!(fn_value.get_string_attribute(AttributeLoc::Return, "my_key"), Some(string_attribute));

pub fn set_param_alignment(self, param_index: u32, alignment: u32)

pub fn as_global_value(self) -> GlobalValue<'ctx>

Gets the GlobalValue version of this FunctionValue. This allows you to further inspect its global properties or even convert it to a PointerValue.

pub fn set_subprogram(self, subprogram: DISubprogram<'ctx>)

Set the debug info descriptor

pub fn get_subprogram(self) -> Option<DISubprogram<'ctx>>

Get the debug info descriptor

pub fn set_section(self, section: &str)

Set the section to which this function should belong