Trait inkwell::values::AggregateValue

pub trait AggregateValue<'ctx>: BasicValue<'ctx> {
    fn as_aggregate_value_enum(&self) -> AggregateValueEnum<'ctx> { ... }

    fn const_extract_value(&self, indexes: &mut [u32]) -> BasicValueEnum<'ctx> { ... }

    fn const_insert_value<BV: BasicValue<'ctx>>(
        &self, 
        value: BV, 
        indexes: &mut [u32]
    ) -> BasicValueEnum<'ctx> { ... }
}

Represents an aggregate value, built on top of other values.

Provided methods

fn as_aggregate_value_enum(&self) -> AggregateValueEnum<'ctx>

Returns an enum containing a typed version of the AggregateValue.

fn const_extract_value(&self, indexes: &mut [u32]) -> BasicValueEnum<'ctx>

fn const_insert_value<BV: BasicValue<'ctx>>(
&self,
value: BV,
indexes: &mut [u32]
) -> BasicValueEnum<'ctx>

Loading content...