Trait inkwell::values::BasicValue

pub trait BasicValue<'ctx>: AnyValue<'ctx> {
    fn as_basic_value_enum(&self) -> BasicValueEnum<'ctx> { ... }

    fn as_instruction_value(&self) -> Option<InstructionValue<'ctx>> { ... }

    fn get_first_use(&self) -> Option<BasicValueUse<'_>> { ... }

    fn set_name(&self, name: &str) { ... }
}

Represents a basic value, which can be used both by itself, or in an AggregateValue.

Provided methods

fn as_basic_value_enum(&self) -> BasicValueEnum<'ctx>

Returns an enum containing a typed version of the BasicValue.

fn as_instruction_value(&self) -> Option<InstructionValue<'ctx>>

Most BasicValues are the byproduct of an instruction and so are convertable into an InstructionValue

fn get_first_use(&self) -> Option<BasicValueUse<'_>>

fn set_name(&self, name: &str)

Sets the name of a BasicValue. If the value is a constant, this is a noop.